Replace fireball material with external shader module
